+++ test/PHPUnit/connecteur/cpp/CPPImporterFactureTest.php
@@ -26,7 +26,9 @@ class CPPImporterFactureTest extends ExtensionCppTestCase
 
         $importerFacture = new CPPImporterFacture($this->getObjectInstancier());
         $importerFacture->setConnecteurId('PortailFacture', $id_ce);
-        $chorusProImportCreationService = $this->getObjectInstancier()->getInstance(ChorusProImportCreationService::class);
+        $chorusProImportCreationService = $this->getObjectInstancier()->getInstance(
+            ChorusProImportCreationService::class
+        );
         $chorusProImportCreationService->setChorusProConfigService(self::ID_E_COL, self::ID_U_ADMIN, $id_ce);
         $this->assertEquals('facture-cpp', $chorusProImportCreationService->getFluxName());
     }
@@ -47,7 +49,9 @@ class CPPImporterFactureTest extends ExtensionCppTestCase
 
         $importerFacture = new CPPImporterFacture($this->getObjectInstancier());
         $importerFacture->setConnecteurId('PortailFacture', $id_ce);
-        $chorusProImportCreationService = $this->getObjectInstancier()->getInstance(ChorusProImportCreationService::class);
+        $chorusProImportCreationService = $this->getObjectInstancier()->getInstance(
+            ChorusProImportCreationService::class
+        );
         $chorusProImportCreationService->setChorusProConfigService(self::ID_E_COL, self::ID_U_ADMIN, $id_ce);
         $this->assertEquals('facture-cpp', $chorusProImportCreationService->getFluxName());
     }
@@ -61,14 +65,18 @@ class CPPImporterFactureTest extends ExtensionCppTestCase
     public function whenConnectorAssociatedWithStatutFactureCpp()
     {
         $this->expectException(Exception::class);
-        $this->expectExceptionMessage("Le connecteur n'est associé à aucun type de dossier pour la récupération des factures...");
+        $this->expectExceptionMessage(
+            "Le connecteur n'est associé à aucun type de dossier pour la récupération des factures..."
+        );
         $result = $this->createConnector('cpp', 'CPP');
         $id_ce = $result['id_ce'];
         $this->associateFluxWithConnector($id_ce, 'statut-facture-cpp', 'PortailFacture');
 
         $importerFacture = new CPPImporterFacture($this->getObjectInstancier());
         $importerFacture->setConnecteurId('PortailFacture', $id_ce);
-        $chorusProImportCreationService = $this->getObjectInstancier()->getInstance(ChorusProImportCreationService::class);
+        $chorusProImportCreationService = $this->getObjectInstancier()->getInstance(
+            ChorusProImportCreationService::class
+        );
         $chorusProImportCreationService->setChorusProConfigService(self::ID_E_COL, self::ID_U_ADMIN, $id_ce);
         $this->assertEquals('facture-cpp', $chorusProImportCreationService->getFluxName());
     }
@@ -78,16 +86,20 @@ class CPPImporterFactureTest extends ExtensionCppTestCase
      */
     public function testCreerFactureSansConnecteurParametrage()
     {
-        $this->mockCPPWrapper(self::DATE_MIN_STATUT_COURANT_OLD, self::DATE_MIN_STATUT_COURANT_OLD);
+        $this->mockCPPWrapper(
+            self::DATE_MIN_STATUT_COURANT_OLD,
+            self::DATE_MIN_STATUT_COURANT_OLD
+        );
         $id_ce = $this->createCppConnector("facture-cpp");
 
         $this->triggerActionOnConnector($id_ce, 'import-facture');
 
         /** @var cpp $cpp */
         $cpp = $this->getConnecteurFactory()->getConnecteurById($id_ce);
-        $depuisLe = $cpp->getDateDepuisLe();
 
-        $last_message_expected = 'Récupération des factures ayant changé de statut depuis le ' . $depuisLe . ' et synchronisation des factures déja présentes:<br/>';
+        $last_message_expected = 'Récupération des factures ayant changé de statut entre le ' .
+            $cpp->getDateDepuisLe() . ' et le ' . $cpp->getDateJusquAu() .
+            ' et synchronisation des factures déja présentes:<br/>';
         $last_message_expected .= '"Création du document" pour 2 facture(s): 4100169, 3346947, <br/>';
         $this->assertLastMessage($last_message_expected);
     }
@@ -97,9 +109,20 @@ class CPPImporterFactureTest extends ExtensionCppTestCase
      */
     public function testImportFactureNominalRecu()
     {
-        $this->setMinDateStatutCourant('recu', self::DATE_MIN_STATUT_COURANT_OLD, ChorusProImportUtilService::TYPE_INTEGRATION_CPP_VALEUR);
-        $this->setMinDateStatutCourant('travaux', self::DATE_MIN_STATUT_COURANT_YOUNG, ChorusProImportUtilService::TYPE_INTEGRATION_CPP_TRAVAUX_VALEUR);
-        $this->mockCPPWrapper(self::DATE_MIN_STATUT_COURANT_OLD, self::DATE_MIN_STATUT_COURANT_YOUNG);
+        $this->setMinDateStatutCourant(
+            'recu',
+            self::DATE_MIN_STATUT_COURANT_OLD,
+            ChorusProImportUtilService::TYPE_INTEGRATION_CPP_VALEUR
+        );
+        $this->setMinDateStatutCourant(
+            'travaux',
+            self::DATE_MIN_STATUT_COURANT_YOUNG,
+            ChorusProImportUtilService::TYPE_INTEGRATION_CPP_TRAVAUX_VALEUR
+        );
+        $this->mockCPPWrapper(
+            self::DATE_MIN_STATUT_COURANT_OLD,
+            self::DATE_MIN_STATUT_COURANT_YOUNG
+        );
         $id_ce = $this->prepareFluxFactureCpp();
         $this->triggerActionOnConnector($id_ce, 'import-facture');
 
@@ -110,7 +133,10 @@ class CPPImporterFactureTest extends ExtensionCppTestCase
         $donneesFormulaire = $this->getDonneesFormulaireFactory()->get($list_facture[0]['id_d']);
 
         $type_integration_libelle = $donneesFormulaire->getFormulaire()->getField('type_integration')->getSelect();
-        $this->assertEquals(ChorusProImportUtilService::TYPE_INTEGRATION_CPP_VALEUR, $type_integration_libelle[$donneesFormulaire->get('type_integration')]);
+        $this->assertEquals(
+            ChorusProImportUtilService::TYPE_INTEGRATION_CPP_VALEUR,
+            $type_integration_libelle[$donneesFormulaire->get('type_integration')]
+        );
 
         $this->assertFileEquals(
             self::FICHIER_PIVOT,
@@ -123,9 +149,20 @@ class CPPImporterFactureTest extends ExtensionCppTestCase
      */
     public function testImportFactureNominalTravaux()
     {
-        $this->setMinDateStatutCourant('recu', self::DATE_MIN_STATUT_COURANT_YOUNG, ChorusProImportUtilService::TYPE_INTEGRATION_CPP_VALEUR);
-        $this->setMinDateStatutCourant('travaux', self::DATE_MIN_STATUT_COURANT_OLD, ChorusProImportUtilService::TYPE_INTEGRATION_CPP_TRAVAUX_VALEUR);
-        $this->mockCPPWrapper(self::DATE_MIN_STATUT_COURANT_YOUNG, self::DATE_MIN_STATUT_COURANT_OLD);
+        $this->setMinDateStatutCourant(
+            'recu',
+            self::DATE_MIN_STATUT_COURANT_YOUNG,
+            ChorusProImportUtilService::TYPE_INTEGRATION_CPP_VALEUR
+        );
+        $this->setMinDateStatutCourant(
+            'travaux',
+            self::DATE_MIN_STATUT_COURANT_OLD,
+            ChorusProImportUtilService::TYPE_INTEGRATION_CPP_TRAVAUX_VALEUR
+        );
+        $this->mockCPPWrapper(
+            self::DATE_MIN_STATUT_COURANT_YOUNG,
+            self::DATE_MIN_STATUT_COURANT_OLD
+        );
         $id_ce = $this->prepareFluxFactureCpp();
         $this->triggerActionOnConnector($id_ce, 'import-facture');
 
@@ -136,7 +173,10 @@ class CPPImporterFactureTest extends ExtensionCppTestCase
         $donneesFormulaire = $this->getDonneesFormulaireFactory()->get($list_facture[0]['id_d']);
 
         $type_integration_libelle = $donneesFormulaire->getFormulaire()->getField('type_integration')->getSelect();
-        $this->assertEquals(ChorusProImportUtilService::TYPE_INTEGRATION_CPP_TRAVAUX_VALEUR, $type_integration_libelle[$donneesFormulaire->get('type_integration')]);
+        $this->assertEquals(
+            ChorusProImportUtilService::TYPE_INTEGRATION_CPP_TRAVAUX_VALEUR,
+            $type_integration_libelle[$donneesFormulaire->get('type_integration')]
+        );
 
         $this->assertFileEquals(
             self::FICHIER_PIVOT,
@@ -149,7 +189,10 @@ class CPPImporterFactureTest extends ExtensionCppTestCase
      */
     public function testNoImportFactureWhenMinDateDepotIsTooYoung()
     {
-        $this->mockCPPWrapper(self::DATE_MIN_STATUT_COURANT_YOUNG, self::DATE_MIN_STATUT_COURANT_YOUNG);
+        $this->mockCPPWrapper(
+            self::DATE_MIN_STATUT_COURANT_YOUNG,
+            self::DATE_MIN_STATUT_COURANT_YOUNG
+        );
         $id_ce = $this->prepareFluxFactureCpp();
         $this->triggerActionOnConnector($id_ce, 'import-facture');
 
@@ -164,7 +207,10 @@ class CPPImporterFactureTest extends ExtensionCppTestCase
      */
     public function testImportFactureWhenMinDateDepotIsOld()
     {
-        $this->mockCPPWrapper(self::DATE_MIN_STATUT_COURANT_OLD, self::DATE_MIN_STATUT_COURANT_OLD);
+        $this->mockCPPWrapper(
+            self::DATE_MIN_STATUT_COURANT_OLD,
+            self::DATE_MIN_STATUT_COURANT_OLD
+        );
         $id_ce = $this->prepareFluxFactureCpp();
         $this->triggerActionOnConnector($id_ce, 'import-facture');
 
@@ -179,16 +225,22 @@ class CPPImporterFactureTest extends ExtensionCppTestCase
      */
     public function testNoCreateFactureRecuWhenStatutFinal()
     {
-        $this->mockCPPWrapper(self::DATE_MIN_STATUT_COURANT_OLD, self::DATE_MIN_STATUT_COURANT_OLD, 'MISE_EN_PAIEMENT');
+        $this->mockCPPWrapper(
+            self::DATE_MIN_STATUT_COURANT_OLD,
+            self::DATE_MIN_STATUT_COURANT_OLD,
+            'MISE_EN_PAIEMENT'
+        );
         $id_ce = $this->createCppConnector("facture-cpp");
 
         $this->triggerActionOnConnector($id_ce, 'import-facture');
 
         /** @var cpp $cpp */
         $cpp = $this->getConnecteurFactory()->getConnecteurById($id_ce);
-        $depuisLe = $cpp->getDateDepuisLe();
 
-        $last_message_expected = 'Récupération des factures ayant changé de statut depuis le ' . $depuisLe . ' et synchronisation des factures déja présentes:<br/>';
+        $last_message_expected =
+            'Récupération des factures ayant changé de statut entre le ' .
+            $cpp->getDateDepuisLe() . ' et le ' . $cpp->getDateJusquAu() .
+            ' et synchronisation des factures déja présentes:<br/>';
         $last_message_expected .= '"Création du document" pour 1 facture(s): 4100169, <br/>';
         $this->assertLastMessage($last_message_expected);
     }
@@ -231,15 +283,26 @@ class CPPImporterFactureTest extends ExtensionCppTestCase
      * @param $min_date_depot_travaux
      * @param string $statut_facture
      */
-    private function mockCPPWrapper($min_date_depot_recu, $min_date_depot_travaux, $statut_facture = 'MISE_A_DISPOSITION')
-    {
+    private function mockCPPWrapper(
+        $min_date_depot_recu,
+        $min_date_depot_travaux,
+        $statut_facture = 'MISE_A_DISPOSITION'
+    ) {
         $cppWrapper = $this->getMockBuilder(CPPWrapper::class)
             ->disableOriginalConstructor()
             ->getMock();
-        $cppWrapper->expects($this->any())->method('rechercheFactureParRecipiendaire')->willReturn($this->getrechercheFactureParRecipiendaire($min_date_depot_recu, $statut_facture));
-        $cppWrapper->expects($this->any())->method('rechercheFactureTravaux')->willReturn($this->getrechercheFactureTravaux($min_date_depot_travaux));
-        $cppWrapper->expects($this->any())->method('telechargerGroupeFacture')->willReturn(file_get_contents(self::FICHIER_PIVOT));
-        $cppWrapper->expects($this->any())->method('consulterHistoriqueFacture')->willReturn($this->getConsulterHistoriqueFacture("2018-04-19 11:16", "MISE_A_DISPOSITION"));
+        $cppWrapper->expects($this->any())
+            ->method('rechercheFactureParRecipiendaire')
+            ->willReturn($this->getrechercheFactureParRecipiendaire($min_date_depot_recu, $statut_facture));
+        $cppWrapper->expects($this->any())
+            ->method('rechercheFactureTravaux')
+            ->willReturn($this->getrechercheFactureTravaux($min_date_depot_travaux));
+        $cppWrapper->expects($this->any())
+            ->method('telechargerGroupeFacture')
+            ->willReturn(file_get_contents(self::FICHIER_PIVOT));
+        $cppWrapper->expects($this->any())
+            ->method('consulterHistoriqueFacture')
+            ->willReturn($this->getConsulterHistoriqueFacture("2018-04-19 11:16", "MISE_A_DISPOSITION"));
 
         $cppWrapperFactory = $this->getMockBuilder(CPPWrapperFactory::class)
             ->disableOriginalConstructor()